ar71xx: Allow port-to-port forward on vlan disable
This fix https://dev.archive.openwrt.org/ticket/11143

Sometimes we need to communicate with a vlan-tagged-upstream.
For example:
We disable vlan on the switch0 and use eth0.5 eth0.9 ... to
communicate with vlan5/vlan9 tagged streams
But when vlan diable, the switch port-to-port is isolate, this
should not be the case.

Signed-off-by: Chen Minqiang <ptpt52@gmail.com>